Move RemoteClient into it's own file.
