+++ api/local.go
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2023, Dana Burkart <dana.burkart@gmail.com>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+package fossil
+
